% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.sf}
\alias{plot.sf}
\title{Plot simple features simply}
\usage{
\method{plot}{sf}(x, ...)
}
\arguments{
\item{x}{sf object (the data.frame one)}

\item{...}{arguments passed to [sf::plot_sfc]}
}
\value{
used for its side effects
}
\description{
Overrides 'sf::plot.sf' and wraps the call to 'plot(st_geometry(x))'.
}
\details{
This can be used when developing code or checking things, so that plots
aren't always faceted by all attributes, coloured by them, or leaving
the plot unusable for subsequent additional drawing.
}
\examples{
nc <- read_sf(system.file("shape/nc.shp", package="sf"))
## all we've changed is the plot command
plot(nc)
plot(nc[sample(1:nrow(nc), 10), ], col = rainbow(10), add = TRUE)
}
